+++ packages/embedded-client/src/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Client library for interacting with embedded Cerbos policy decision points from server-side Node.js and browser-based applications.
+ *
+ * @packageDocumentation
+ */
+
+export * from "./client";
+export * from "./options";